created utils file with functions
